fix overbudget

+++ src/modules/garment-purchasing/purchase-order-external/template/purchase-order-external-item.js
@@ -41,8 +41,9 @@ export class PurchaseOrderItem {
   checkIsOverBudget() {
     if (this.context.context.options.checkOverBudget) {
       var totalDealPrice = ((this.data.dealQuantity * this.price * this.kurs.rate) + this.data.budgetUsed).toFixed(4);
-      var totalBudget=parseInt(this.data.totalBudget.toFixed(4));
+      var totalBudget=parseFloat(this.data.totalBudget.toFixed(4));
       //console.log(totalDealPrice + " >"+this.data.totalBudget.toFixed(4));
+      
       if (totalDealPrice > totalBudget) {
         this.data.isOverBudget = true;
         // console.log(totalDealPrice + " >"+totalBudget);
@@ -54,6 +55,7 @@ export class PurchaseOrderItem {
       
     }
   }
+
 
   updatePrice() {
     this.data.priceBeforeTax = this.price;
